Add RandomWorkerImg
