+++ src/components/WokerCard/RandomWorkerImg/style.ts
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+
+export const WorkerImgBox = styled.div`
+  background-color: #f5f6f8;
+  border-radius: 10px;
+  display: flex;
+  padding: 1.6875rem 2.5rem 0rem 2.6875rem;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+`;
