+++ src/components/WokerCard/style.ts
@@ -5,16 +5,6 @@ export const WorkerCardContainer = styled.div`
   height: 14.0625rem;
 `;
 
-export const WorkerImgBox = styled.div`
-  background-color: #f5f6f8;
-  border-radius: 10px;
-  display: flex;
-  padding: 1.6875rem 2.5rem 0rem 2.6875rem;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-`;
-
 export const WorkerInfoHead = styled.div`
   font-size: 1.125rem;
   font-weight: 600;
